refactor: move DemonstrationsModel config from training to controller module
- Relocate DemonstrationsModel schema from training.schemas to controller.schemas
- Update imports in DemonstrationsDataset to reflect new location
- Add demonstrations field to CONTROLLER_USER_CONFIG
- Fix syntax error in CONTROLLER_USER_CONFIG (missing comma)
- Add debug output for configuration errors in train command
